refactor: remove extra logic from mixins
